Backends now created using a factory
The concept of 'data-stores' has been renamed to 'backends', which can
now be registered to be created automatically using a factory method in
src/backends/backend.[h|cpp]. The filesystem has been modified to store
this backends into an array which can be indexed using the backend type.
At the moment, two backends (DRAM and NVRAM-NVML) can be chosen by the
user. What remains to be done is specify exactly to which backend a file
needs to go. Also, the code for the NVRAM-NVML backend is just a stub at
the moment.