Preload files sequentially into NVRAM using NVML
Files specified by the user in the configuration file can now be
preloaded into NVRAM using the pmem_map_file function from the NVML
library. The full file is preloaded, and the filesystem verifies whether
the file is already available in DRAM or NVRAM before serving requests
to it from the underlying filesystem.